requirements.txt
streamlit
supabase
pandas
plotly
python-dotenv
streamlit-lottie
requests